Extract FakeRoomMembersIterator. Cannot make it implement the interface instead of extend Rust class RoomMembersIterator for now.
