+++ libraries/matrix/impl/src/test/kotlin/io/element/android/libraries/matrix/impl/sdk/FakeRoomMembersIterator.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * Please see LICENSE in the repository root for full details.
+ */
+
+package io.element.android.libraries.matrix.impl.sdk
+
+import org.matrix.rustcomponents.sdk.NoPointer
+import org.matrix.rustcomponents.sdk.RoomMember
+import org.matrix.rustcomponents.sdk.RoomMembersIterator
+
+// Note: need to extend RoomMembersIterator until RoomInterface.members() and
+// RoomInterface.membersNoSync() returns RoomMembersIteratorInterface
+class FakeRoomMembersIterator(
+    private var members: List<RoomMember>? = null
+) : RoomMembersIterator(NoPointer) {
+    override fun len(): UInt {
+        return members?.size?.toUInt() ?: 0u
+    }
+
+    override fun nextChunk(chunkSize: UInt): List<RoomMember>? {
+        if (members?.isEmpty() == true) {
+            return null
+        }
+        return members?.let {
+            val result = it.take(chunkSize.toInt())
+            members = it.subList(result.size, it.size)
+            result
+        }
+    }
+}
